Split crypto.c into sha.c and aes.c
Likewise for crypto.h into sha.h and sha.c, of course.
