+++ source/decryptor/aes.h
@@ -37,33 +37,12 @@
 #define AES_CNT_TITLEKEY_ENCRYPT_MODE (AES_CBC_ENCRYPT_MODE | AES_CNT_INPUT_ORDER | AES_CNT_OUTPUT_ORDER | AES_CNT_INPUT_ENDIAN | AES_CNT_OUTPUT_ENDIAN)
 
 
-#define REG_SHACNT      ((volatile uint32_t*)0x1000A000)
-#define REG_SHABLKCNT   ((volatile uint32_t*)0x1000A004)
-#define REG_SHAHASH     ((volatile uint32_t*)0x1000A040)
-#define REG_SHAINFIFO   ((volatile uint32_t*)0x1000A080)
-
-#define SHA_CNT_STATE           0x00000003
-#define SHA_CNT_OUTPUT_ENDIAN   0x00000008
-#define SHA_CNT_MODE            0x00000030
-#define SHA_CNT_ENABLE          0x00010000
-#define SHA_CNT_ACTIVE          0x00020000
-
-#define SHA_HASH_READY          0x00000000
-#define SHA_NORMAL_ROUND        0x00000001
-#define SHA_FINAL_ROUND         0x00000002
-
-#define SHA256_MODE             0
-#define SHA224_MODE             0x00000010
-#define SHA1_MODE               0x00000020
-
-
-void add_ctr(void* ctr, u32 carry);
-
 void setup_aeskeyX(u8 keyslot, void* keyx);
 void setup_aeskeyY(u8 keyslot, void* keyy);
 void setup_aeskey(u8 keyslot, void* keyy);
 void use_aeskey(u32 keyno);
 void set_ctr(void* iv);
+void add_ctr(void* ctr, u32 carry);
 void aes_decrypt(void* inbuf, void* outbuf, size_t size, u32 mode);
 void aes_fifos(void* inbuf, void* outbuf, size_t blocks);
 void set_aeswrfifo(u32 value);
@@ -72,7 +51,3 @@ u32 aes_getwritecount();
 u32 aes_getreadcount();
 u32 aescnt_checkwrite();
 u32 aescnt_checkread();
-
-void sha_init(u32 mode);
-void sha_update(const void* src, u32 size);
-void sha_get(void* res);
